Added Access Control

+++ app/Http/Controllers/PostController.php
@@ -8,6 +8,7 @@ use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
 use Illuminate\Http\Response;
 use Illuminate\Routing\Redirector;
+use Illuminate\Support\Facades\Storage;
 use Illuminate\View\View;
 
 class PostController extends Controller
@@ -44,15 +45,36 @@ class PostController extends Controller
     public function store(Request $request)
     {
         $this->validate($request, [
-           'title' => 'required',
-           'body' => 'required'
+            'title' => 'required',
+            'body' => 'required',
+            'cover_image' => 'image|nullable|max:1999'
         ]);
+
+        // File Upload Handler
+        if ($request->hasFile('cover_image'))
+        {
+            // Get filename with the extension
+            $filenameWithExt = $request->file('cover_image')->getClientOriginalName();
+            // Get just the file name
+            $filename = pathinfo($filenameWithExt, PATHINFO_FILENAME);
+            // Get just the extension
+            $extension = $request->file('cover_image')->getClientOriginalExtension();
+            // Filename to store
+            $fileNameToStore = $filename . '_' . time() . '.' . $extension;
+            // Upload Image
+            $path = $request->file('cover_image')->storeAs('public/cover_images', $fileNameToStore);
+        }
+        else
+        {
+            $fileNameToStore = 'noimage.jpg';
+        }
 
         // Create Post
         $post = new Post();
         $post->title = $request->input('title');
         $post->body = $request->input('body');
         $post->user_id = auth()->user()->id;
+        $post->cover_image = $fileNameToStore;
         $post->save();
 
         return redirect('/posts')->with('success', 'Created a new post.');
@@ -78,7 +100,14 @@ class PostController extends Controller
      */
     public function edit($id)
     {
-        return view('posts.edit')->with('post', Post::findOrFail($id));
+        $post = Post::findORFail($id);
+
+        // Check for User Match
+        if (auth()->user()->id !== $post->user_id)
+        {
+            return redirect('/posts')->with('error', 'Nice try, Gringo. You cannot edit someone else\'s post. -5 points for Gryffindor.');
+        }
+        return view('posts.edit')->with('post', $post);
     }
 
     /**
@@ -92,13 +121,45 @@ class PostController extends Controller
     {
         $this->validate($request, [
             'title' => 'required',
-            'body' => 'required'
+            'body' => 'required',
+            'cover_image' => 'image|nullable|max:1999'
         ]);
 
-        // Create Post
+        // File Upload Handler
+        if ($request->hasFile('cover_image'))
+        {
+            // Get filename with the extension
+            $filenameWithExt = $request->file('cover_image')->getClientOriginalName();
+            // Get just the file name
+            $filename = pathinfo($filenameWithExt, PATHINFO_FILENAME);
+            // Get just the extension
+            $extension = $request->file('cover_image')->getClientOriginalExtension();
+            // Filename to store
+            $fileNameToStore = $filename . '_' . time() . '.' . $extension;
+            // Upload Image
+            $path = $request->file('cover_image')->storeAs('public/cover_images', $fileNameToStore);
+        }
+        else
+        {
+            $fileNameToStore = 'noimage.jpg';
+        }
+
+        // Update Post
         $post = Post::find($id);
         $post->title = $request->input('title');
         $post->body = $request->input('body');
+
+        // Checks if the post has a cover image, if yes, it automatically posts again its own cover image if
+        // the user doesn't upload a new cover image.
+        // if the user did upload again, it deletes the previous file and replaces it with the new one
+        if ($request->hasFile('cover_image'))
+        {
+            if ($post->cover_image != 'no_image.png')
+            {
+                Storage::delete('public/cover_images/'.$post->cover_image);
+            }
+            $post->cover_image = $fileNameToStore;
+        }
         $post->save();
 
         return redirect('/posts')->with('success', 'Post updated.');
@@ -113,7 +174,25 @@ class PostController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+
+        // Check for User Match
+        if (auth()->user()->id !== $post->user_id)
+        {
+            return redirect('/posts')->with('error', 'Nice try, Gringo. You cannot delete someone else\'s post. -5 points for Gryffindor.');
+        }
+
+        if ($post->cover_image != 'noimage.jpg')
+        {
+            // Delete Image
+            Storage::delete('public/cover_images/' . $post->cover_image);
+        }
+
         $post->delete();
         return redirect('/posts')->with('error', 'Post deleted.');
+    }
+
+    public function __construct()
+    {
+        $this->middleware('auth', ['except' => ['index', 'show']]);
     }
 }
